requirements.txt
Django==4.2.7
python-decouple==3.8
psycopg-binary==3.2.12
psycopg==3.2.12
django-extensions==3.2.3
djangorestframework==3.14.0
PyJWT>=2.8.0
shapely>=2.0.4
django-widget-tweaks==1.5.0
requests>=2.31.0
django-redis>=5.4.0

# Tests & CI (Sprint 08)
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
